+++ lib/class/BaseComponent/index.d.ts
@@ -6,7 +6,7 @@ export default class BaseComponent<T = {}> {
     protected classes: TClasses;
     protected containerElement: HTMLElement | null;
     private _state?;
-    private initiated;
+    protected initiated: boolean;
     protected get state(): T;
     protected set state(value: T | undefined);
     constructor(props: IConstructorBaseProps, initState?: T);