+++ lib/type.d.ts
@@ -27,7 +27,6 @@ export declare enum EEVentName {
     ADAPTATION = "adaptation",
     RATE_CHANGE = "ratechange",
     SETTING_PANEL_VISIBLE = "settingPanelVisible",
-    SETTING_PANEL_BLUR = "settingPanelBlur",
     VARIANT_CHANGED = "variantchanged",
     ABR_STATUS_CHANGED = "abrstatuschanged",
     VOLUME_CHANGE = "volumechange",
@@ -74,7 +73,6 @@ export interface SmListeners {
     [EEVentName.SETTING_PANEL_VISIBLE]: (event: EEVentName.SETTING_PANEL_VISIBLE, data: {
         visible: boolean;
     }) => void;
-    [EEVentName.SETTING_PANEL_BLUR]: (event: EEVentName.SETTING_PANEL_BLUR, data?: any) => void;
     [EEVentName.VOLUME_CHANGE]: (event: EEVentName.VOLUME_CHANGE, data: {
         [key: string]: any;
     }) => void;